[fix] procedural canvas circle algorithm
